Update ButtonPanel component
