+++ src/components/App.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import './App.css';
 
+import ButtonPanel from './ButtonPanel';
+
 function App() {
   return (
     <div className="App">
-      <h1> Welcome to React Calculator </h1>
+      <ButtonPanel/>
     </div>
   );
 }
